+++ packages/fast-colors/src/range.ts
